Move eviction trait and strategies in a separate module
