+++ vm/src/handlers/lib/mod.rs
@@ -1,6 +1,7 @@
 use common::client_commands::Message;
 
 pub mod ecall;
+pub mod evict;
 pub mod outsourced_mem;
 pub mod vapp;
